Remove generics from Trace, TraceLine
This simplifies code and makes possible to unify Reporter and TaskReporter (also Report and TaskReport).
In future it may be required to do some `instanceof` checks in Reporter implementations that needs to detect which subclass of Trace they received however this will be much simpler/cleaner than polluting Reporter/Report/Trace/TraceLine with generics.